refactor: create PortRange.swift
