+++ IVPNClient/Models/PortRange.swift
@@ -1,0 +1,46 @@
+//
+//  PortRange.swift
+//  IVPN iOS app
+//  https://github.com/ivpn/ios-app
+//
+//  Created by Juraj Hilje on 2022-07-15.
+//  Copyright (c) 2022 Privatus Limited.
+//
+//  This file is part of the IVPN iOS app.
+//
+//  The IVPN iOS app is free software: you can redistribute it and/or
+//  modify it under the terms of the GNU General Public License as published by the Free
+//  Software Foundation, either version 3 of the License, or (at your option) any later version.
+//
+//  The IVPN iOS app is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+//  or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+//  details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with the IVPN iOS app. If not, see <https://www.gnu.org/licenses/>.
+//
+
+import Foundation
+
+struct PortRange {
+    var tunnelType: String
+    var protocolType: String
+    var ranges: [String]
+    
+    var storeKey: String {
+        if tunnelType == "OpenVPN" {
+            if protocolType == "TCP" {
+                return UserDefaults.Key.openvpnTcpCustomPort
+            } else {
+                return UserDefaults.Key.openvpnUdpCustomPort
+            }
+        }
+        
+        return UserDefaults.Key.wireguardCustomPort
+    }
+    
+    func save(port: Int) {
+        UserDefaults.standard.set(port, forKey: storeKey)
+    }
+}
